make block realtime resizable
This is changing Block so the allocated memory and the actual accessible
memory are two different things. The block will just expose a clear
function and a raw_buffer_mut method, access should now only go via
views so that callers only see the data they are supposed to see.
This also adds some functions to initialize blocks from pointers.